+++ experiments/mirrors_content_analysis_2026_04_24/analysis/length_compression_analysis/__init__.py
@@ -1,0 +1,6 @@
+from experiments.mirrors_content_analysis_2026_04_24.analysis.length_compression_analysis.main import (
+    LengthCompressionAnalyzer,
+    main,
+)
+
+__all__ = ["LengthCompressionAnalyzer", "main"]
