+++ experiments/mirrors_content_analysis_2026_04_24/analysis/readability_complexity_analysis/__init__.py
@@ -1,0 +1,6 @@
+from experiments.mirrors_content_analysis_2026_04_24.analysis.readability_complexity_analysis.main import (
+    ReadabilityComplexityAnalyzer,
+    main,
+)
+
+__all__ = ["ReadabilityComplexityAnalyzer", "main"]
